pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["HarshavardhanaP96 <pkhrao3112.harsha@gmail.com>"]
readme = "README.md"

packages = [
    { include = "config" }
]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "3.16.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.2"
black = "^25.9.0"
isort = "^6.0.1"
mypy = "^1.18.2"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pre-commit = "^4.3.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
select = ["E", "F", "W", "I", "C", "N"]

[tool.mypy]
python_version = 3.11
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
